requirements.txt
langchain
langchain-openai
openai
pyyaml
python-dotenv
aiohttp
pandas
tqdm